more specific software requirements

+++ sentinel.py
@@ -1,25 +1,33 @@
 #!/usr/bin/python
 
 # SENTINEL
+# A USB rocket launcher face-tracking solution
 # For Linux and Windows
-# Tested on Python 2.7
 #
 # Hardware requirements:
 # - Dream Cheeky brand USB rocket launcher (tested with Thunder model, should also work with Storm)
 # - small webcam attached to USB rocket launcher, in /dev/video0
 #
 # Software requirements (Linux):
+# - Python 2.7, 32-bit
 # - libusb (in Ubuntu/Debian, apt-get install libusb-dev)
 # - PyUSB 1.0 (https://github.com/walac/pyusb)
+# - NumPy (in Ubuntu/Debian, apt-get install python-numpy)
 # - OpenCV Python bindings (in Ubuntu/Debian, apt-get install python-opencv)
-# - PIL, if not already installed
+# - PIL (in Ubuntu/Debian, apt-get install python-imaging)
 # - streamer (in Ubuntu/Debian, apt-get install streamer)
 #
 # Software requirements (Windows):
-# - libusb (http://www.craftedge.com/products/libusb.html) and create USB_Missile_Launcher.inf
+# - Python 2.7, 32-bit
+# - libusb (http://sourceforge.net/projects/libusb-win32/files/)
+#     - After installing, plug in USB rocket launcher, launch <libusb path>\bin\inf-wizard.exe,
+#       and create and run an INF driver file for the USB rocket launcher using the wizard
 # - PyUSB 1.0 (https://github.com/walac/pyusb)
-# - OpenCV Python bindings (http://docs.opencv.org/doc/tutorials/introduction/windows_install/windows_install.html#windows-installation)
-# - PIL (from http://www.lfd.uci.edu/~gohlke/pythonlibs/)
+# - NumPy (http://www.lfd.uci.edu/~gohlke/pythonlibs/#numpy)
+# - OpenCV Python bindings (http://sourceforge.net/projects/opencvlibrary/files/opencv-win/2.3.1/OpenCV-2.3.1-win-superpack.exe/download)
+#     - After installing, copy the contents of <opencv path>\build\python\2.7 (it should contain cv.py and cv2.pyd)
+#       to c:\Python27\Lib
+# - PIL (http://www.lfd.uci.edu/~gohlke/pythonlibs/#pil)
 
 import os
 import sys
@@ -85,13 +93,16 @@ class Turret():
 
 class Camera():
    def dispose(self):
-      os.system("killall display")
+      if os.name == 'posix':
+         os.system("killall display")
 
    def capture(self, img_file):
       if os.name == 'posix':
          os.system("streamer -c /dev/video0 -b 16 -o " + img_file)
+         # generates 320x240 greyscale jpeg
       else:
          os.system("CommandCam")
+         # generates 640x480 color bitmap
 
    def face_detect(self, img_file, haar_file, out_file):
       hc = cv.Load(haar_file)
